refactor: draft 관련 DTO 클래스명 변경
